refactor: 디렉토리 반영 후 클래스 이동
