+++ src/main/java/lotto/domain/LottoPurchaseAmount.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class LottoPurchaseAmount {
     private static final int LOTTO_PRICE = 1000;
