+++ src/main/java/lotto/view/InputView.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
+import lotto.domain.Lotto;
 
 import java.util.Arrays;
 import java.util.List;
