+++ src/test/java/lotto/LottoLottoPurchaseAmountTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.LottoPurchaseAmount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
