+++ src/test/java/lotto/LottoPaperTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.LottoPaper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
